Updated DiskAccessLibrary

+++ DiskAccessLibrary/FileSystems/NTFS/ClusterUsageBitmap.cs
@@ -1,4 +1,4 @@
-/* Copyright (C) 2014 Tal Aloni <tal.aloni.il@gmail.com>. All rights reserved.
+/* Copyright (C) 2014-2018 Tal Aloni <tal.aloni.il@gmail.com>. All rights reserved.
  * 
  * You can redistribute this program and/or modify it under the terms of
  * the GNU Lesser Public License as published by the Free Software Foundation,
@@ -7,7 +7,6 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using Utilities;
 
 namespace DiskAccessLibrary.FileSystems.NTFS
@@ -207,8 +206,8 @@ namespace DiskAccessLibrary.FileSystems.NTFS
 
         public static void UpdateClusterStatus(byte[] bitmap, ulong clusterIndexInBitmap, bool isUsed)
         {
-            ulong byteOffset = (ulong)clusterIndexInBitmap / 8;
-            int bitOffset = (int)clusterIndexInBitmap % 8;
+            ulong byteOffset = clusterIndexInBitmap / 8;
+            int bitOffset = (int)(clusterIndexInBitmap % 8);
             if (isUsed)
             {
                 bitmap[byteOffset] |= (byte)(0x01 << bitOffset);
